+++ examples/tracing.rs
@@ -5,10 +5,10 @@ use anyhow::{anyhow, Result};
 use futures::StreamExt;
 use scylla::batch::Batch;
 use scylla::execution::iterator::RowIterator;
+use scylla::execution::tracing::TracingInfo;
 use scylla::statement::{
     prepared_statement::PreparedStatement, query::Query, Consistency, SerialConsistency,
 };
-use scylla::tracing::TracingInfo;
 use scylla::QueryResult;
 use scylla::{Session, SessionBuilder};
 use std::env;
