+++ scylla/src/execution/mod.rs
@@ -4,6 +4,7 @@ pub mod iterator;
 pub mod load_balancing;
 pub mod retries;
 pub mod speculative_execution;
+pub mod tracing;
 
 pub(crate) use execution_profile::ExecutionProfileInner;
 pub use execution_profile::{ExecutionProfile, ExecutionProfileBuilder, ExecutionProfileHandle};
